Add identity
A `-generate` parameter to generate a text version of a secret key.

This key can and should be kept a\on a postit note :-)
Just don't store somewhere insecure. It's your future identity.

This can be set as an environment variable:

```bash
unset HISTFILE
 export GO_MYSPACE_CLIENT_IDENTITY=FooBarABCDEFbase58
```

or specified on the command line:

```bash
./go-myspace-client -identity FooBarABCDEFbase58
```

The first is the best.