Refactor P2P to be a struct
This should greatly simply both use and configuration of P2P
functionality.

It only takes a Kademlia DHT instance or libp2p Options as aparameters.

The DHT Instance can be nil, but if you want to confifure routing or set
ServerMode that is possible and makes sense.

As for libp2p options, these are required to enable relay services and such.