Refactor ConectinGater and Discovery
The discovery process had become way to complex and I
hadn't realised how important it is for the host to constantly
have a connection to a slew of other hosts to function properly.

It's not enough to know each other, kademlia works best, when
you have a reasonable amount of anonymous peers. The gater is
still there for the future, but it's not in use.